+++ src/query/handleErrorMessage.ts
@@ -14,5 +14,23 @@ export function handleErrorMessage(error: ApiError | any, translate: TFunction<"
     const message = error.body?.message || error.statusText;
     if (!code || code === 500) return { message: translate("somethingWentWrong"), type: "error" };
     else if (code === 401) return { message: translate("sessionExpired"), type: "warning" };
-    else return { message: translate(message in en.error ? error.body.message : "somethingWentWrong"), type: "error" };
+    else if (error instanceof Error) {
+        const stringifiedError = error.toString();
+        if (stringifiedError.includes("cover storage")) {
+            return {
+                message: translate("notEnoughBalanceForStorage"),
+                type: "error",
+            };
+        } else if (stringifiedError.includes("enough balance")) {
+            return {
+                message: translate("notEnoughBalance"),
+                type: "error",
+            };
+        } else {
+            return {
+                message: translate("somethingWentWrong"),
+                type: "error",
+            };
+        }
+    } else return { message: translate(message in en.error ? error.body.message : "somethingWentWrong"), type: "error" };
 }